Additional dock functionality and project refactor
The first thing I wanted to do was refactor the file structure for the
project as well as begin calling it *dock* instead of *fucking*. Makes
the project feel a bit more serious, lol.

As far as the actual program, I added a couple of functions:
- `startup()`: creates a temporary directory for persistent storage used
during `dock`
- `cleanup()`: removes the persistant storage once the function is done. I
plan to include a debug option to keep the persistant storage if desired
- `kill_all()`: kills all of the docker containers that are running. If the
container is in a stopped state, this function will not remove the container.
Planning on fixing this by adding error handling to the function, but for
now it is ok.

- Filesystem restructure
- Function addition: startup, cleanup and kill_all