Update kill commands
The kill_all function needed updates and the kill function was written.

`kill_all()` was given a bit of intel to check if a container is already
running or not. This changes whether or not to deliver the `docker kill`
or not.

`kill()` was added to allow being able to provide either the containerID
or the containerName in order to kill and remove the docker container.

List of Changes
- `kill_all()` updates
- `kill()` creation