fix: Got SP interception workflow working
Going forward, the plan is to just generate a mock Service Provider from the cli
and have the user configure the IdP url to then configure. It will be important
to allow for further SP configuration for all cases, but the idea will be that
the user can get sent to a localhost call, which requires auth from their
configured IdP. Once the IdP authenticates, the service can then intercept the
SAMLResponse token to assume the AWS role. Will it work with a more official IdP?
I don't know honestly...